feat: add tool to patch HydraBATS role
This tool is a way of "fixing" the `HydraBATSRole` so that the same Hydra Stack can be used to run tests in several pipelines.

The tool Adds a pipelineId to the list of "bats.amazon.com:entity" in the assume role policy of HydraBATSRole.
This is needed to share a hydra stack between different pipelines when using CDK.
See: https://sim.amazon.com/issues/T2-10912 (LPT does allow you to override the pipeline list in the role)